Update setup.sh
The existing *setup.sh* script no longer worked as expected. (See AB#172643 and other bugs.) After publishing the functions, the Azure portal would not list the functions. This was because the existing Zip files contained more code than was necessary for the newer implementation of Azure functions. With that in mind, this new script extracts the contents of the Zip files and builds a deployment-ready Zip package with just the necessary files, while leaving the original Zip files with all of the original code in this repo.